+++ src/main/java/exception/ResourceTypeNotFoundException.java
@@ -1,0 +1,12 @@
+package exception;
+
+public class ResourceTypeNotFoundException extends RuntimeException{
+
+    public ResourceTypeNotFoundException() {
+
+    }
+
+    public ResourceTypeNotFoundException(String message) {
+        super(message);
+    }
+}
