--- src/main/java/http/ResourceTypeNotFoundException.java
@@ -1,12 +1,0 @@
-package http;
-
-public class ResourceTypeNotFoundException extends RuntimeException{
-
-    ResourceTypeNotFoundException() {
-
-    }
-
-    ResourceTypeNotFoundException(String message) {
-        super(message);
-    }
-}
